fix: move visualization selector  from light effects to LED control panel
