+++ src/modules/sbstudio/plugin/overlays/__init__.py
@@ -1,13 +1,13 @@
 from .base import Overlay
 from .formation_order import FormationOrderOverlay
-from .light_effect import LightEffectOverlay
+from .leds import LEDsOverlay
 from .pyro import PyroOverlay
 from .safety_check import SafetyCheckOverlay
 
 __all__ = (
     "Overlay",
     "FormationOrderOverlay",
-    "LightEffectOverlay",
+    "LEDsOverlay",
     "PyroOverlay",
     "SafetyCheckOverlay",
 )
